Simplify and clean event handling
